Consolidate Google Firewall port checking and add case of vars
